Acabat sprint 2

+++ app/Actions/Jetstream/DeleteUser.php
@@ -37,8 +37,10 @@ class DeleteUser implements DeletesUsers
     {
         $user->teams()->detach();
 
-        $user->ownedTeams->each(function (Team $team) {
-            $this->deletesTeams->delete($team);
+        $user->ownedTeams->each(function ($team) {
+            if ($team instanceof Team) {
+                $this->deletesTeams->delete($team);
+            }
         });
     }
 }
